Add impersonation support to libcurl
Add impersonation support to libcurl. A new function,
curl_easy_impersonate() was added. It sets the needed options on the
Curl_easy handle.

In addition, it is possible to supply the CURL_IMPERSONATE environment
variable with the impersonation target (e.g. 'ff91esr'), and libcurl
will automatically call curl_easy_impersonate() on any new handle. This
allows seamless integreation with change the code of the app that uses
libcurl.

For now, CURL_IMPERSONATE does not fully work when curl in launched
from the command line, as the command line tool overrides some of our
options (the SSL version for example).