Make the TLS ALPS extension configurable
The ALPS extension was previously added to libcurl-impersonate to
impersonate Chrome, but was hardcoded to be on. This makes it
configurable via CURLOPT_SSL_ENABLE_ALPS and the '--alps' command line
option.

In addition, the NPN extension was previously disabled by commenting out
the relevant code. However, there is CURLOPT_SSL_ENABLE_NPN and
'--no-npn' which do the same thing. Therefore the commented out block
was restored to work as in the upstream curl.