Add support for compiling NSS statically on macOS
Statically linking with NSS is a total mess and completely
undocumented. There are 20+ .a files to link with, and their linking
order matters. The main reference for this commit is a Mozilla Rust code
responsible for statically linking NSS:

https://github.com/mozilla/application-services/blob/b2690fd2e4cc3e8e10b6868ab0de8b79c89d3a93/components/support/rc_crypto/nss/nss_b
uild_common/src/lib.rs#L94

Even with that in hand some hacking of linker flags is needed to make
the curl's autoconf linking tests pass.

In addition, the configure script will now exit with error if it fails
linking with NSS, preventing the common case of curl being compiled
without SSL support.