Change order of curl-impersonate's header building
curl-impersonate merges a list of 'base headers' with user supplied
headers. However, this was done after the 'Host' header is determined,
causing some edge cases to appear in which an incorrect 'Host' header is
sent of a connection is re-used. To fix it, the order of the calls was
changed.