Immitate Firefox HTTP2 stream priorities
Firefox uses an elaborate scheme of http/2 streams to split the load
between html/js/css/img resources.  It builds a tree of streams with
different weights (priorities) by default and communicates this to the
server. With this commit this behavior is immitated in order to improve
the http/2 impersonation.