Use optional CURL_IMPERSONATE_HEADERS env variable
The CURL_IMPERSONATE_HEADERS is an optional environment variable which
is inspected whenever the CURL_IMPERSONATE environment variable is set.
When set to "no", the default list of built-in headers will not be
used when impersonating. This can give greater control to the user of
the used HTTP headers content and order in case these are needed.